Decompile more functions
