+++ config/saturn/zero_user_syms.txt
@@ -33,8 +33,6 @@ _DMA_ScuGetStatus = 0x06023A60;
 _DMA_ScuIntFunc0 = 0x06023AB0;
 _DMA_ScuIntFunc1 = 0x06023B54;
 _DMA_ScuIntFunc2 = 0x06023C00;
-_get_imask = 0x06023CAC;
-_set_imask = 0x06023CE4;
 _func_06024088 = 0x06024088;
 _SCL_Memcpyw = 0x06024B20;
 _SCL_PriIntProc = 0x06024C20;
@@ -62,6 +60,8 @@ _MTH_Div = 0x0602F570;
 _fcos = 0x0602FC20;
 _fsin = 0x0602FC26;
 _func_06030184 = 0x06030184;
+_SYS_GetMac = 0x06030A4C;
+_SYS_SetMac = 0x06030A54;
 _func_06030df0 = 0x06030DF0;
 ___ashiftrt_r4_24 = 0x06030E2A;
 ___sdivsi3 = 0x06030E6C;
