+++ tools/builds/saturn.py
@@ -102,6 +102,7 @@ lib_srcs = [
     'src/saturn/lib/cdc.c',
     'src/saturn/lib/csh.c',
     'src/saturn/lib/dma/dma_scu1.c',
+    'src/saturn/lib/dma/dma_scu0.c',
     'src/saturn/lib/dma/dma.c',
     'src/saturn/lib/gfs.c',
     'src/saturn/lib/mth.c',
@@ -180,6 +181,7 @@ multi_objs = {
         'build/saturn/lib/cdc.o',
         'build/saturn/lib/csh.o',
         'build/saturn/lib/dma/dma_scu1.o',
+        'build/saturn/lib/dma/dma_scu0.o',
         'build/saturn/lib/dma/dma.o',
         'build/saturn/lib/gfs.o',
         'build/saturn/lib/mth.o',
